Create Madrid target win
Real Madrid's Rodrygo, Bellingham on target in win over Braga

Real Madrid's Rodrygo and Jude Bellingham scored in each half to secure a 2-1 win at Sporting Braga and keep their side three points clear atop Champions League Group C on Tuesday.

It was an affair that won't live long the memory on a cold rainy day in northern Portugal in which the Spanish giants dominated from the start and opened the scoring in the 16th minute with a close-range strike by Rodrygo.

Jet-heeled Vinícius Júnior flew up the left wing and put the ball on a plate for his fellow Brazilian to tap in with a sliding first touch from inside the six-yard box.

Wasteful Madrid missed several chances to extend their lead mainly through Vinícius, who was often caught offside playing on the shoulder of an extremely high Braga defensive line.

But they managed to score again in the 61st when Rodrygo beat the offside trap and found Vinicius ghosting behind the defence on the left.

He collected the through ball and passed to Bellingham who, from the edge of the box, slotted an effortless curling strike just inside the far post.

It was Bellingham's 10th goal in 11 games in all competitions this season as he became only the second Real Madrid player to score in each of his first three Champions League games after Christian Karembeu in 1998.

https://club.vexanium.com/post/guarda-taylor-swift-the-eras-tour-2023-film-completo-intero-online-gratuito-6538d4d288ec8b7ceddb10fa
https://club.vexanium.com/post/guarda-l-esorcista---il-credente-2023-streaming-in-italiano-intero-online-g--6538d56f717de803527a38ab
https://club.vexanium.com/post/guarda-paw-patrol---il-super-film-2023-film-completo-streaming-ita-online-g--6538d5df34a43808c17702e9
https://club.vexanium.com/post/guarda-saw-x-2023-streaming-ita-film-completo-gratis-gratuito-6538d63ad2dd13118075b505
https://club.vexanium.com/post/guarda-the-nun-ii-2023-film-completo-streaming-in-italiano-altadefinizione-6538d696f3f2aa814928282b
https://club.vexanium.com/post/guarda-five-nights-at-freddy-s-2023-streaming-ita-gratis-film-completo-alta--6538d6ebdf6e4039127afc76
https://club.vexanium.com/post/guarda-killers-of-the-flower-moon-2023-streaming-ita-online-film-completo-g--6538d72ed0779f229708f55a
https://club.vexanium.com/post/guarda-the-marvels-2023-film-completo-online-streaming-italiano-gratuito-6538d786f3f2aaa104282a35
https://club.vexanium.com/post/guarda-the-creator-2023-streaming-in-italiano-film-completo-gratis-gratuito-6538d7cf42e7967920bcb7d9
https://club.vexanium.com/post/guarda-anatomia-di-una-caduta-2023-film-completo-ita-online-altadefinizione--6538d89188ec8b17f8db138f
https://club.vexanium.com/post/assista-taylor-swift-the-eras-tour-2023-filme-completo-online-dublado-grati--6538d974d0779f296308f60e
https://club.vexanium.com/post/assistir-o-exorcista-o-devoto-2023-dublado-e-legendado-filme-completo-em-po--6538d9d4adf1b802311367d5
https://club.vexanium.com/post/assista-patrulha-canina-o-filme-superpoderoso-2023-filme-completo-dublado-e--6538da1cadf1b883ba1367e0
https://club.vexanium.com/post/assistir-jogos-mortais-x-2023-filme-completo-dublado-gratis-em-portugues-6538da864a856b4dccd8bbd5
https://club.vexanium.com/post/assistir-a-freira-2-2023-dublado-e-legendado-gratis-filme-completo-6538dadcd0779f7e4508f6b3
https://club.vexanium.com/post/assista-five-nights-at-freddy-s---o-pesadelo-sem-fim-2023-dublado-e-legenda--6538db1fd2d2863c7cd3888d
https://club.vexanium.com/post/assistir-assassinos-da-lua-das-flores-2023-dublado-e-legendado-gratis-em-po--6538db6ed2d2863bbad38922
https://club.vexanium.com/post/assista-as-marvels-2023-filme-completo-dublado-gratis-legendado-6538dbbf8d31d1f15bf21df6
https://club.vexanium.com/post/assista-resistencia-2023-filme-completo-gratis-dublado-em-portugues-6538dc0889d70f1472fc04bd
https://club.vexanium.com/post/assistir-anatomie-d-une-chute-2023-filme-completo-dublado-em-portugues-grat--6538dc6cb0ed21def5ca97cc
https://club.vexanium.com/post/mira-taylor-swift-the-eras-tour-2023-pelicula-completa-en-linea-gratis-lati--6538dd460b581840d53be19c
https://club.vexanium.com/post/ver-el-exorcista-creyentes-2023-pelicula-completa-en-espanol-gratis-latino-6538dd979552962805ea0ed7
https://club.vexanium.com/post/repelis-paw-patrol-la-super-pelicula-2023-pelicula-completa-en-espanol-onli--6538dddc98b587d40dcc514d
https://club.vexanium.com/post/ver-saw-x-el-juego-del-miedo-2023-pelicula-completa-en-linea-cuevana-latino-6538de4459587f87100ca786
https://club.vexanium.com/post/repelis-la-monja-ii-2023-pelicula-completa-en-linea-latino-online-cuevana-6538de898b28ce7b124af757
https://club.vexanium.com/post/repelis-five-nights-at-freddy-s-2023-pelicula-completa-en-espanol-latino-y---6538ded359587f001e0ca836
https://club.vexanium.com/post/ver-los-asesinos-de-la-luna-2023-pelicula-completa-en-espanol-gratis-y-cast--6538df2b95529675a9ea11ae
https://club.vexanium.com/post/repelis-the-marvels-2023-pelicula-completa-en-linea-latino-y-subtitulado-6538df78e5e60ea091d4d019
https://club.vexanium.com/post/ver-resistencia-2023-pelicula-completa-en-espanol-latino-y-castellano-6538dfc66c27e64563764a6d
https://club.vexanium.com/post/ver-anatomy-of-a-fall-2023-pelicula-completa-en-linea-gratis-y-subtitulado-6538e0121968662461ecb283
https://club.vexanium.com/post/nezd-taylor-swift-the-eras-tour-2023-teljes-film-magyarul-online-letoltes-v--6538e3ff2a1cca1f202c882a
https://club.vexanium.com/post/nezd-az-ordoguzo-a-hivo-2023-teljes-film-magyar-online-filmek-videa-6538e44496807603318885d7
https://club.vexanium.com/post/nezd-mancs-orjarat-a-szuperfilm-2023-teljes-film-magyarul-letoltes-online-i--6538e486e8e86bf74e97c466
https://club.vexanium.com/post/nezd-furesz-10-2023-teljes-film-magyarul-online-filmek-magyar-alcim-letolte--6538e4c183e71315fbb03dc9
https://club.vexanium.com/post/nezd-az-apaca-ii-2023-teljes-film-magyar-alcim-letoltes-online-videa-6538e4fcc3c52eab4926d5cc
https://club.vexanium.com/post/nezd-ot-ejjel-freddy-pizzazojaban-2023-teljes-film-magyarul-online-filmek-m--6538e53e589b839a1f5d492b
https://club.vexanium.com/post/nezd-megfojtott-viragok-2023-teljes-film-magyar-alcim-ingyen-online-letolte--6538e57f894fe35f1310bbb4
https://club.vexanium.com/post/nezd-marvelek-2023-teljes-film-magyar-alcim-letoltes-online-videa-6538e5c7e8e86b3a0d97c6c8
https://club.vexanium.com/post/nezd-az-alkoto-2023-teljes-film-magyar-online-filmek-letoltes-videa-6538e6035e7f6b38afc7eb6e
https://club.vexanium.com/post/nezd-egy-zuhanas-anatomiaja-2023-teljes-film-magyarul-online-filmek-magyar---6538e6493f6e7d554721b4b4

Braga did pull one back when Álvaro Djaló took advantage of a defensive mistake in the 63rd to score in the area, but Madrid held on despite late pressure for their third win in three games thanks to two brilliant save by Kepa Arrizabalaga.

"We had to suffer more than expected in the end but it's normal to suffer away from home, especially in the Champions League," Vinicius told Movistar Plus. "The important thing is that we won. We played well, although it's true that we could have scored more goals.

"The summary is that we won all three group games so far. Let's keep this run going."

Carlo Ancelotti's side is top on nine points ahead of Napoli, who eked out a 1-0 win at bottom side Union Berlin with the Germans yet to earn a point. Braga are in third place on three points.

Real Madrid coach Carlo Ancelotti believes Vinícius Júnior will soon win a Ballon d'Or award.

The 23 year-old is on the 2023 shortlist for the award, that hasn't been won by a Brazilian player since Kaka in 2007.

Asked if he thinks there is a Brazilian at Madrid that could win the top individual award one day, Ancelotti told TNT Sports after his team's 2-1 Champions League win at Braga : "Yes, Vinícius will win it soon."