Create Man Utd defeat
Erik ten Hag has said Manchester United's heavy derby defeat to Manchester City was one of his worst days as manager at Old Trafford.

City cruised to a 3-0 win on Sunday thanks to two goals from Erling Haaland, including a first-half penalty, and another from Phil Foden.

Ten Hag has suffered some dismal results as United manager, but asked whether the latest defeat to Pep Guardiola's side made it one of his worst days in the job, he said: "Yes."

He added: "Of course, it is disappointing, but last year we had many highlights. When you lose a derby in the way we lose, that is disappointing.

"First half we had a very good game plan and the execution was also very good. It was toe-to-toe first and it was very similar, but then the penalty changed the moment. Then, in the second half we chose to become more offensive and it is 2-0 too quickly. From that point on it was a difficult game."

United have now lost five of their first 10 league games for the first time since 1986 and sit eighth in the Premier League table, already 11 points behind leaders Tottenham.

https://sketchfab.com/3d-models/taylor-swift-the-eras-tour-film-completo-ita-37a24d6ad11e4c1a8b4c22fe9852c839
https://sketchfab.com/3d-models/totally-killer-film-completo-ita-dd72180e42c24300a2f5032601a3e0a1
https://sketchfab.com/3d-models/once-upon-a-studio-film-completo-ita-b4ff07859a684797a1249d2643555a5f
https://sketchfab.com/3d-models/saw-x-film-completo-ita-ee45d2a5d5df4a6f819b9ff2e5bb335f
https://sketchfab.com/3d-models/the-burial-il-caso-okeefe-film-completo-ita-6d0394b75e29446a9495d60f2cfd3f7f
https://sketchfab.com/3d-models/five-nights-at-freddys-film-completo-ita-db04d136ccee4a298fd4aa4ff7283f0f
https://sketchfab.com/3d-models/killers-of-the-flower-moon-film-completo-ita-97f3cd25982d4aaaa75a3c94182ff4dc
https://sketchfab.com/3d-models/the-marvels-film-completo-ita-beb1fe6a27ac4e2ab1f0bb95ee508084
https://sketchfab.com/3d-models/the-creator-film-completo-ita-758b41cbfc764d95b9484731fbfeb6df
https://sketchfab.com/3d-models/reptile-film-completo-ita-8190d0a375b14cad98f2119bfd0d4fc2
https://sketchfab.com/3d-models/taylor-swift-the-eras-tour-pelicula-completa-cbc46c1290de4af2a971eca6becaa8fc
https://sketchfab.com/3d-models/dulces-y-sangrientos-16-pelicula-completa-8038fcbcde924479a025911e1b47c08a
https://sketchfab.com/3d-models/habia-una-vez-un-estudio-pelicula-completa-ce0bf3c36cf6498d8315faab46063626
https://sketchfab.com/3d-models/saw-x-el-juego-del-miedo-pelicula-completa-cd87090cba98444c86934c93adb312b3
https://sketchfab.com/3d-models/enterrando-una-ambicion-pelicula-completa-91c1015e30184b10a6677ff4c12bacee
https://sketchfab.com/3d-models/five-nights-at-freddys-pelicula-completa-ea702f6c5ade4e6b91a06c08abc692e6
https://sketchfab.com/3d-models/los-asesinos-de-la-luna-pelicula-completa-2c21e8d554824654bdff4909aa051ac2
https://sketchfab.com/3d-models/the-marvels-pelicula-completa-066b4e13714949c681220628fdbc29a8
https://sketchfab.com/3d-models/resistencia-pelicula-completa-761db1fba3b042499d962284c3ecf911
https://sketchfab.com/3d-models/reptiles-pelicula-completa-latino-29e49f70f5a44dc298c7d750ffa4f58c
https://sketchfab.com/3d-models/nezd-taylor-swift-the-eras-tour-teljes-film-5d88a05d92a04486b27d93b84ae48eb6
https://sketchfab.com/3d-models/nezd-gyilkos-a-multbol-teljes-film-d96ff0d59e314590bcd5b68ae8da20f7
https://sketchfab.com/3d-models/nezd-volt-egyszer-egy-studio-teljes-film-fa7c9a9ab1c34e6f8ec0ab3c10a62d74
https://sketchfab.com/3d-models/nezd-furesz-10-teljes-film-598eaea46d7e417689910e357592e3cf
https://sketchfab.com/3d-models/nezd-a-temetes-teljes-film-83b5b1133271490b9255c369cb083da8
https://sketchfab.com/3d-models/nezd-ot-ejjel-freddy-pizzazojaban-teljes-film-43d62de1941c407f9fee4f1816d4479f
https://sketchfab.com/3d-models/nezd-megfojtott-viragok-teljes-film-bf438270bec748ac8ff55bd68efc8d44
https://sketchfab.com/3d-models/nezd-marvelek-teljes-film-49aeacf656c0453687a9c4a867cd438d
https://sketchfab.com/3d-models/nezd-az-alkoto-teljes-film-32fb19919e2846c9a9f3d0177765dce3
https://sketchfab.com/3d-models/nezd-hidegver-teljes-film-9d6f9a6772ac4db694359027a1367fae
https://sketchfab.com/3d-models/taylor-swift-the-eras-tour-filme-completo-86f0378ba69c418f809aaac3f2c6177a
https://sketchfab.com/3d-models/dezesseis-facadas-filme-completo-bec6df9fb4e14241b4639ff27f5250ee
https://sketchfab.com/3d-models/era-uma-vez-um-estudio-filme-completo-12c016360c774b53ad23a49e45ca0aa2
https://sketchfab.com/3d-models/jogos-mortais-x-filme-completo-2a2355bb24434092a427d97a189f0758
https://sketchfab.com/3d-models/o-proprio-enterro-filme-completo-5fd43f49bff041eda45a085d3f7f9e34
https://sketchfab.com/3d-models/freelance-filme-completo-469f0e3539824a6f8824b351069939bf
https://sketchfab.com/3d-models/assassinos-da-lua-das-flores-filme-completo-dcefc3f6f2204ed281a3ca6694ba304c
https://sketchfab.com/3d-models/as-marvels-filme-completo-87cdaa88f3a54a7094c4d1f506e5335d
https://sketchfab.com/3d-models/resistencia-filme-completo-f897fa0ed42e4a4799ff07c89979eb16
https://sketchfab.com/3d-models/camaleoes-filme-completo-dfb63797370d42abb4518ca6e7a31bcc

It is now seven defeats from 14 games in all competitions this season but Ten Hag said he is confident his team are making progress.

"The three games before we won and the spirit is very good," Ten Hag said. "I think we are on the way up. The start was difficult but now we are on the way up. When injuries are getting back we will be getting stronger. We have to be patient but I am happy with some injuries coming back and then our side will be stronger."

Afterwards, Ten Hag said he stood by his decision to drop Raphaël Varane and Sergio Reguilón to the bench in favour of Jonny Evans and Victor Lindelöf while also defending his decision to replace striker Rasmus Højlund midway through the second half after the substitution was greeted with loud boos from the United fans inside Old Trafford.

"I have to protect Rasmus Højlund and I have to protect the team," Ten Hag added.

"He is putting so much effort in pressing, going in the transitions, going the long ways, fighting a tough opposition.

"He is not used to it, three games in a week, so I have to protect him and the team, to bring some energy in. We have bench players, Mason Mount and [Alejandro] Garnacho, we know he can change games."